refactor: split Home component
