+++ app/client/src/components/home/types.ts
@@ -1,0 +1,11 @@
+export type Account = {
+  id: string;
+  userName: string;
+  displayName: string;
+  avatarInitial: string;
+  publicKey?: string;
+  followers?: string[];
+  following?: string[];
+};
+
+export const isDataUrl = (str: string) => str.startsWith("data:image/");
